+++ tests/test_job.c
@@ -69,7 +69,7 @@ void
 __CUT__job_test_hash_free()
 {
     job j;
-    unsigned long long int jid = 83;
+    uint64_t jid = 83;
 
     j = make_job_with_id(0, 0, 1, 0, default_tube, jid);
     job_free(j);
@@ -81,7 +81,7 @@ void
 __CUT__job_test_hash_free_next()
 {
     job a, b;
-    unsigned long long int aid = 97, bid = 12386;
+    uint64_t aid = 97, bid = 12386;
 
     b = make_job_with_id(0, 0, 1, 0, default_tube, bid);
     a = make_job_with_id(0, 0, 1, 0, default_tube, aid);
